Refactor unlock system: Introduce UnlockType and ChallengeType, update related classes and methods

+++ src/main/java/com/rogueliteplugin/challenge/ChallengeState.java
@@ -2,16 +2,16 @@ package com.rogueliteplugin.challenge;
 
 public class ChallengeState {
     private final Challenge challenge;
-    private long progress;
-    private final long goal;
+    private int progress;
+    private final int goal;
 
-    public ChallengeState(Challenge challenge, long goal) {
+    public ChallengeState(Challenge challenge, int goal) {
         this.challenge = challenge;
         this.goal = goal;
         this.progress = 0;
     }
 
-    public void increment(long amount) {
+    public void increment(int amount) {
         progress = Math.min(goal, progress + amount);
     }
 
@@ -27,11 +27,11 @@ public class ChallengeState {
         return challenge;
     }
 
-    public long getProgress() {
+    public int getProgress() {
         return progress;
     }
 
-    public long getGoal() {
+    public int getGoal() {
         return goal;
     }
 }
